Rephrased reference to regular expressions
This section of your text refers to pattern matching in .gitignore, which doesn't match patterns as POSIX regular expressions, but rather uses [its own pattern matching rules](https://git-scm.com/docs/gitignore). These rules include treating the pattern as a shell glob (using the metacharacters *, ? and [ to match file names in the command line). You don't go into globbing in much detail, except in the section titled "Get Wild" (under "Working with Unix"). The reference to regular expressions here is confusing. It may be better to describe it using the language from that section ("Get Wild"): "patterns that use wild cards", rather than to say "regular expressions", and then qualify that as "anything that you can use with `ls`".  The suggested change in phrasing should remind readers of that subsection, which is helpful, since the `*` shell glob metacharacter is exactly what you describe in that subsection.